fix blog action
