+++ next.config.ts
@@ -5,7 +5,7 @@ const removeImports = require("next-remove-imports");
 const nextConfig: NextConfig = {
     experimental: {
         serverActions: {
-            bodySizeLimit: "2mb",
+            bodySizeLimit: "4mb",
         },
         turbo: {
             rules: {
