+++ src/config/images.ts
@@ -1,0 +1,2 @@
+export const blogImageUrl = "https://mnkroicwfgayhyxomjqt.supabase.co/storage/v1/object/public/blogs//blog.png";
+export const projectImageUrl = "https://mnkroicwfgayhyxomjqt.supabase.co/storage/v1/object/public/blogs//project.png";
